virt-api: imporve the client API usage for the pod eviction handling
This will replace the usage of GET and UPDATE client methods with a
PATCH method. In addition, instead of fetching the VMI and the Pod
objects directly from the API server, use shared informer cache.

recently a high flakiness was observed from the migration e2e tests
related to live migration upon pod eviction request. It appears from
the virt-api logs that the client failed to update the server with the
the relevant VMI status fields.

Signed-off-by: Igor Bezukh <ibezukh@redhat.com>