e2e: config_test: adjust helper function for libvmi usage
libvmi VM factory doesn't configure namespace for the VMI
the namespace is defined later on byt the functions that use
the Create API call.

Therefore namespace could not be read from the VMI object
amd the util.TestNamespaceDefault can be safely used, since
its value is calculated in runtime to support parallel runs.

Signed-off-by: Igor Bezukh <ibezukh@redhat.com>