Test: Introduce condition matcher
This matcher should show more useful information
in case a condition is missing or the status is false.

Previously asserting conditions was done with
for loop and it produced output:

```
Expected false to be true
```
Now we can see:

```
Expected
    <here will be slice of conditions>

to find condition of type 'Paused' and status 'Fasle' but got 'True'
```

This hopefuly provides better picture without looking into code.

Signed-off-by: L. Pivarc <lpivarc@redhat.com>