cpu hotplug: fix improper reject of VM configuration
in order to reduce the burden from the VM owner, the maxSockets field is optional.
If it's not specified neither in the VM spec nor in the kubevirt CR, it can be
automatically calculated from requested sockets with a ratio. In that case, there
is no sense in rejecting amount of sockets that is exceeding the current calculated maximum.
Instead, stage the change. It will be applied after a power cycle, with a new calculated
maximum.

Signed-off-by: Igor Bezukh <ibezukh@redhat.com>