virt-handler: force ignore VMI sync while hotplug migration takes place
CPU and memory hotplug actions lead to changes in the VM spec,
which immediately propogated to the VMI object. On virt-handler side
the syncVMI will be called since the VMI spec has changed.
The problem is that SyncVMI will carry the updated CPU/memory topology.
On the live-migration source virt-launcher it can cause failure because
the VMI-to-domain conversion is using the source compute container cpuset,
which has less resources than what is in the updated VMI.

Signed-off-by: Igor Bezukh <ibezukh@redhat.com>